+++ pkg/apitype/policy.go
@@ -14,6 +14,8 @@
 
 package apitype
 
+import "encoding/json"
+
 // DefaultPolicyGroup is the name of the default Policy Group for organizations.
 const DefaultPolicyGroup = "default-policy-group"
 
@@ -29,6 +31,10 @@ type CreatePolicyPackRequest struct {
 
 	// A pretty name for the Policy Pack that is supplied by the package.
 	DisplayName string `json:"displayName"`
+
+	// VersionTag is the semantic version of the Policy Pack. One a version is published, it
+	// cannot never be republished. Older clients will not have a version tag.
+	VersionTag string `json:"versionTag,omitempty"`
 
 	// The Policies outline the specific Policies in the package, and are derived
 	// from the package by the CLI.
@@ -52,11 +58,19 @@ type RequiredPolicy struct {
 	// The version of the required Policy Pack.
 	Version int `json:"version"`
 
+	// The version tag of the required Policy Pack.
+	VersionTag string `json:"versionTag"`
+
 	// The pretty name of the required Policy Pack.
 	DisplayName string `json:"displayName"`
 
 	// Where the Policy Pack can be downloaded from.
 	PackLocation string `json:"packLocation,omitempty"`
+
+	// The configuration that is to be passed to the Policy Pack. This is map a of policies
+	// mapped to their configuration. Each individual configuration must comply with the
+	// JSON schema for each Policy within the Policy Pack.
+	Config map[string]*json.RawMessage `json:"config,omitempty"`
 }
 
 // Policy defines the metadata for an individual Policy within a Policy Pack.
@@ -73,7 +87,30 @@ type Policy struct {
 	// Message is the message that will be displayed to end users when they violate
 	// this policy.
 	Message string `json:"message"`
+
+	// The JSON schema for the Policy's configuration.
+	ConfigSchema *PolicyConfigSchema `json:"configSchema,omitempty"`
 }
+
+// PolicyConfigSchema defines the JSON schema of a particular Policy's
+// configuration.
+type PolicyConfigSchema struct {
+	// Config property name to JSON Schema map.
+	Properties map[string]*json.RawMessage `json:"properties,omitempty"`
+	// Required config properties.
+	Required []string `json:"required,omitempty"`
+
+	// Type defines the data type allowed for the schema.
+	Type JSONSchemaType `json:"type"`
+}
+
+// JSONSchemaType in an enum of allowed data types for a schema.
+type JSONSchemaType string
+
+const (
+	// Object is a dictionary.
+	Object JSONSchemaType = "object"
+)
 
 // EnforcementLevel indicates how a policy should be enforced
 type EnforcementLevel string
@@ -93,6 +130,7 @@ type GetPolicyPackResponse struct {
 	Name        string   `json:"name"`
 	DisplayName string   `json:"displayName"`
 	Version     int      `json:"version"`
+	VersionTag  string   `json:"versionTag"`
 	Policies    []Policy `json:"policies"`
 	Applied     bool     `json:"applied"`
 }
@@ -112,8 +150,9 @@ type UpdatePolicyGroupRequest struct {
 	AddStack    *PulumiStackReference `json:"addStack,omitempty"`
 	RemoveStack *PulumiStackReference `json:"removeStack,omitempty"`
 
-	AddPolicyPack    *PolicyPackMetadata `json:"addPolicyPack,omitempty"`
-	RemovePolicyPack *PolicyPackMetadata `json:"removePolicyPack,omitempty"`
+	AddPolicyPack          *PolicyPackMetadata `json:"addPolicyPack,omitempty"`
+	RemovePolicyPack       *PolicyPackMetadata `json:"removePolicyPack,omitempty"`
+	UpdatePolicyPackConfig *PolicyPackMetadata `json:"updatePolicyPackConfig,omitempty"`
 }
 
 // PulumiStackReference contains the StackName and ProjectName of the stack.
@@ -127,6 +166,11 @@ type PolicyPackMetadata struct {
 	Name        string `json:"name"`
 	DisplayName string `json:"displayName"`
 	Version     int    `json:"version"`
+	VersionTag  string `json:"versionTag"`
+
+	// The configuration that is to be passed to the Policy Pack. This
+	// map ties Policies with their configuration.
+	Config map[string]*json.RawMessage `json:"config,omitempty"`
 }
 
 // ListPolicyPacksResponse is the response to list an organization's
@@ -137,9 +181,10 @@ type ListPolicyPacksResponse struct {
 
 // PolicyPackWithVersions details the specifics of a Policy Pack and all its available versions.
 type PolicyPackWithVersions struct {
-	Name        string `json:"name"`
-	DisplayName string `json:"displayName"`
-	Versions    []int  `json:"versions"`
+	Name        string   `json:"name"`
+	DisplayName string   `json:"displayName"`
+	Versions    []int    `json:"versions"`
+	VersionTags []string `json:"versionTags"`
 }
 
 // ListPolicyGroupsResponse lists a summary of the organization's Policy Groups.
@@ -154,4 +199,11 @@ type PolicyGroupSummary struct {
 	IsOrgDefault          bool   `json:"isOrgDefault"`
 	NumStacks             int    `json:"numStacks"`
 	NumEnabledPolicyPacks int    `json:"numEnabledPolicyPacks"`
+}
+
+// GetPolicyPackConfigSchemaResponse is the response that includes the JSON
+// schemas of Policies within a particular Policy Pack.
+type GetPolicyPackConfigSchemaResponse struct {
+	// The JSON schema for each Policy's configuration.
+	ConfigSchema map[string]PolicyConfigSchema `json:"configSchema,omitempty"`
 }